// mat4331-basetalk-crawler/src/common/types/bat-info.type.ts
export interface BatInfo {
  PA: number;
  AB: number;
  R: number;
  H: number;
  HR: number;
  RBI: number;
  BB: number;
  HBP: number;
  SO: number;
  GO: number;
  FO: number;
  NP: number;
  GDP: number;
  LOB: number;
  ABG: number;
  OPS: number;
  LI: number;
  WPA: number;
  RE24: number;
}